Refactor CLI commands & calls to address generation functions
- Modified CLI to take two commands: hot and cold
- Simplified input validation
- Moved code logic for generating multiple addresses into `Wallet` and
  derived classes
