+++ core/src/main/java/com/ashelkov/wallet/Application.java
@@ -4,7 +4,6 @@ import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.beust.jcommander.JCommander;
 import com.beust.jcommander.ParameterException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,19 +13,14 @@ import com.ashelkov.wallet.bip.Coin;
 import com.ashelkov.wallet.bip.address.Bip44Address;
 import com.ashelkov.wallet.bip.wallet.Wallet;
 import com.ashelkov.wallet.bip.wallet.WalletFactory;
-import com.ashelkov.wallet.io.FileUtils;
-import com.ashelkov.wallet.io.Parameters;
+import com.ashelkov.wallet.io.util.FileUtils;
+import com.ashelkov.wallet.io.Params;
 
 public class Application {
 
     private static final Logger logger = LoggerFactory.getLogger(Application.class);
-    private static final Parameters params = Parameters.getInstance();
-    private static final JCommander commander = JCommander.newBuilder().addObject(params).build();
+    private static final Params params = Params.getInstance();
     private static final SecureRandom rng = Utils.getSecureRandomInstance();
-
-    private enum OutputMode {
-        MNEMONIC_ONLY, DEFAULT_ADDRESS, SPECIFIC_ADDRESS, HOT_WALLET
-    }
 
     public static void main(String[] args) {
 
@@ -35,18 +29,16 @@ public class Application {
         //
 
         try {
-            commander.parse(args);
+            params.parseArguments(args);
         } catch (ParameterException e) {
             logger.error(e.getMessage());
             System.exit(1);
         }
 
         if (params.isHelp()) {
-            commander.usage();
+            params.printUsage();
             System.exit(0);
         }
-
-        OutputMode outputMode = validateInput();
 
         //
         // Generate mnemonic
@@ -76,29 +68,23 @@ public class Application {
         //
 
         List<Bip44Address> addresses;
-        int numAddresses = params.getNumAddresses();
 
-        switch (outputMode) {
-            case DEFAULT_ADDRESS -> {
-                addresses = new ArrayList<>(numAddresses);
-                generateDefaultAddresses(addresses, params.getCoin(), seedFromMnemonic, numAddresses);
-            }
-            case SPECIFIC_ADDRESS -> {
-                addresses = new ArrayList<>(1);
-                generateSpecificAddress(addresses, seedFromMnemonic);
-            }
-            case HOT_WALLET -> {
-                addresses = new ArrayList<>(numAddresses * Coin.values().length);
-                generateHotWallet(addresses, seedFromMnemonic, numAddresses);
+        switch (params.getCommand()) {
+            case Params.COMMAND_COLD -> {
 
+                Wallet wallet = WalletFactory.generateWallet(seedFromMnemonic, params.getCoin());
+                addresses = wallet.generateAddresses(
+                    params.getAccount(),
+                    params.getChange(),
+                    params.getIndex(),
+                    params.getNumAddresses());
             }
-            case MNEMONIC_ONLY -> {
-                addresses = new ArrayList<>(0);
+            case Params.COMMAND_HOT -> {
+                addresses = new ArrayList<>(Coin.values().length);
+                generateHotAddresses(addresses, seedFromMnemonic);
             }
-            default -> {
-                // This case can never occur, but the Java compiler cannot detect that this is so
-                addresses = new ArrayList<>(0);
-            }
+            default ->
+                throw new IllegalArgumentException("Unrecognized command");
         }
 
         //
@@ -110,100 +96,10 @@ public class Application {
         FileUtils.saveAddressesToFile(params.getOutputDirectory(), mnemonic, addresses);
     }
 
-    private static OutputMode validateInput() {
-
-        OutputMode result;
-
-        boolean aci = ((params.getAccount() != null) || (params.getChange() != null) || (params.getIndex() != null));
-        boolean coin = params.getCoin() != null;
-        boolean hot = params.isHot();
-        boolean multi = params.getNumAddresses() > 1;
-
-        boolean error = false;
-
-        //
-        // Errors
-        //
-
-        if (aci && !coin) {
-            logger.error(
-                    "Setting any of the BIP44 account, change, or index elements explicitly requires specifying a " +
-                            "specific coin for which to generate an address");
-            error = true;
-
-        } else if (coin && hot) {
-            logger.error("Both address mode and hot wallet mode enabled; disable one and re-run");
-            error = true;
-        }
-
-        if (error) {
-            System.exit(1);
-        }
-
-        //
-        // Warnings
-        //
-
-        if (!coin && !hot) {
-            logger.warn("Neither address mode nor hot wallet mode enabled; generating mnemonic only");
-
-            if (multi) {
-                logger.warn("No addresses to generate; ignoring multi address input");
-            }
-        }
-
-        if (aci && multi) {
-            logger.warn("Creating specific BIP44 address overrides creation of multiple addresses; ignoring multi " +
-                    "address input");
-        }
-
-        //
-        // Determine output mode
-        //
-
-        if (aci) {
-            result = OutputMode.SPECIFIC_ADDRESS;
-        } else if (coin) {
-            result = OutputMode.DEFAULT_ADDRESS;
-        } else if (hot) {
-            result = OutputMode.HOT_WALLET;
-        } else {
-            result = OutputMode.MNEMONIC_ONLY;
-        }
-
-        return result;
-    }
-
-    private static void generateDefaultAddresses(List<Bip44Address> addresses, Coin coin, byte[] seed, int iter) {
-
-        Wallet wallet = WalletFactory.generateWallet(seed, coin);
-
-        for (int i = 0; i < iter; ++i) {
-
-            Bip44Address address = wallet.getDefaultAddress(i);
-            addresses.add(address);
-
-            logger.trace(address.getCoin());
-            logger.trace(address.getPath());
-            logger.trace(address.getAddress());
-        }
-    }
-
-    private static void generateSpecificAddress(List<Bip44Address> addresses, byte[] seed) {
-
-        Wallet wallet = WalletFactory.generateWallet(seed, params.getCoin());
-        Bip44Address address = wallet.getSpecificAddress(params.getAccount(), params.getChange(), params.getIndex());
-        addresses.add(address);
-
-        logger.trace(address.getCoin());
-        logger.trace(address.getPath());
-        logger.trace(address.getAddress());
-    }
-
-    private static void generateHotWallet(List<Bip44Address> addresses, byte[] seed, int iter) {
-
+    private static void generateHotAddresses(List<Bip44Address> addresses, byte[] seed) {
         for (Coin coin : Coin.values()) {
-            generateDefaultAddresses(addresses, coin, seed, iter);
+            Wallet wallet = WalletFactory.generateWallet(seed, coin);
+            addresses.add(wallet.generateDefaultAddresses(1).get(0));
         }
     }
 }
