+++ core/src/main/java/com/ashelkov/wallet/bip/wallet/StellarWallet.java
@@ -1,5 +1,8 @@
 package com.ashelkov.wallet.bip.wallet;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.stellar.sdk.KeyPair;
 
 import com.ashelkov.wallet.bip.util.Bip44Utils;
@@ -22,7 +25,7 @@ public class StellarWallet extends Wallet {
     }
 
     @Override
-    public Bip44Address getSpecificAddress(Integer account, Integer change, Integer addressIndex) {
+    public List<Bip44Address> generateAddresses(Integer account, Integer change, Integer index, int numAddresses) {
 
         if (account == null) {
             logMissing(ACCOUNT);
@@ -31,16 +34,30 @@ public class StellarWallet extends Wallet {
         if (change != null) {
             logWarning(CHANGE, change);
         }
-        if (addressIndex != null) {
-            logWarning(INDEX, addressIndex);
+        if (index != null) {
+            logWarning(INDEX, index);
         }
 
-        return getAddress(account);
+        List<Bip44Address> result = new ArrayList<>(numAddresses);
+
+        for(int i = account; i < (account + numAddresses); ++i) {
+            result.add(getAddress(i));
+        }
+
+        return result;
     }
 
     @Override
-    public Bip44Address getDefaultAddress(int index) {
-        return getAddress(index);
+    public List<Bip44Address> generateDefaultAddresses(int numAddresses) {
+
+        List<Bip44Address> result = new ArrayList<>(numAddresses);
+        int account = 0;
+
+        for(int i = account; i < (account + numAddresses); ++i) {
+            result.add(getAddress(i));
+        }
+
+        return result;
     }
 
     private Bip44Address getAddress(int account) {
